fix suggested user profile
