+++ client/src/App.js
@@ -31,6 +31,7 @@ import ForgotPassword from './pages/recovery/forgotPassword';
 import ResetPassword from './pages/recovery/resetPassword';
 import ConfirmDeletePost from './components/home/ConfirmDeletePost';
 import Reports from './pages/admin/Reports';
+import { SERVER_URL } from './utils/config';
 
 function App() {
   const { auth, status, modal, call, share, theme, deleteModal, report } = useSelector(state => state)
@@ -42,7 +43,7 @@ function App() {
     dispatch(refreshToken())
 
     // create new socket 
-    const socket = io('https://vyviegramserver.onrender.com')
+    const socket = io(SERVER_URL)
     dispatch({type: GLOBALTYPES.SOCKET, payload: socket})
     return () => socket.close()
   },[dispatch])
