+++ client/src/utils/config.js
@@ -1,1 +1,2 @@
 export const BASE_URL = 'https://vyviegramnetwork.onrender.com'
+export const SERVER_URL = 'https://vyviegramserver.onrender.com'
